minor changes to progression event
