+++ src/handlers/instance.rs
@@ -130,7 +130,6 @@ pub async fn create_minecraft_instance(
         async move {
             let (progression_start_event, event_id) = Event::new_progression_event_start(
                 format!("Setting up Minecraft server {instance_name}"),
-                Some(uuid.clone()),
                 Some(10.0),
                 Some(ProgressionStartValue::InstanceCreation {
                     instance_uuid: uuid.clone(),
@@ -291,7 +290,6 @@ pub async fn delete_instance(
         } else {
             let (progression_event_start, event_id) = Event::new_progression_event_start(
                 format!("Deleting instance {}", instance.name().await),
-                Some(uuid.clone()),
                 Some(10.0),
                 None,
                 caused_by,
